refactor(longevity-counters-multidc): Switching from c-s user profile to scylla-bench counters
Cassandra-stress counters and c-s user profile with counters doesn't really work.
There are long staning issues for this problem.
This commit changes the command for longevity-counters-multidc from c-s to scylla-bench.
It also introduce a new way to pre-create keyspace by passing the CREATE KEYSPACE DDL command through the yaml.

The scylla-bench counter_update and counter_read commands are working with iterations instead of duration.
So, I removed the 4d from the longevity name and the test duration is based on the number of iterations.