feature(nemesis): Introduce new type of nemesis sequential calling and AdminPerfImpactMonkey
1. Introudcing a new function call_sequential_disrupt_method.
This function is being used to call a list of nemesis,
sequanitlly, covering all the nemesis in the list.
The function also support multiplying the members of the list and looping over the list.

2. Introducing new "property" for some nemesis: "admin_perf_impact".
The nemesis that are set with this "property" are collected into a list of nemesis
by the class AdminPerfImpactMonkey.
This class is used for performance testing when we would like to measure the impact of
some nemesis on performance.2. Introducing new "property" for some nemesis: "admin_perf_impact".
The nemesis that are set with this "property" are collected into a list of nemesis
by the class AdminPerfImpactMonkey.
This class is used for performance testing when we would like to measure the impact of
some nemesis on performance.