test(longevity-200gb-48h): Changing the nemesis class to Sisyphus
Till now we used a hard coded list called LimitedChaosMonkey to set
the nemesis that run in this longevity.
With this change it will collect all the nemesis that has the
property 'limited' set to true.
Switching to Sisyphus will give us the ability to repeat the exact
same sequence of nemesis if needed by using nemesis_seed.