fix(performance_regression_test): Wait no compactions after write
Since there are ongoing changes with compactions, there are cases
where compactions leftovers are still running after the prepare
workload and affecting the results of READ and MIXED workloads.

It was decided to fix it on the original branch and not advancing
to a new branch version despite the fact it may void some of the
previous results.

(cherry picked from commit 8e78ae9981135e16cd0d73cfb6045e64d0c6ac8f)