refactor(SCHEMA_FAILURE): Change default severity of SCHEMA_FAILURE
SCT detects the regex "Failed to load schema version" and creates
an ERROR event out of it even if the message in scylla is with INFO
severity. this behviour raises issues like
"https://github.com/scylladb/scylladb/issues/19093" which gets P3
priority but continue failing SCT tests over and over.
Hence, this commit suggests to reduce this message severity.